Change component to not "take over" the openshift-monitoring namespace
This commit changes the component to not make the `openshift-monitoring`
namespace object itself managed by Syn ArgoCD.

This change ensures that deactivating the component on a cluster will
not delete the whole OpenShift cluster-monitoring component.

Instead, the component now uses the resource-locker component to add the
desired labels to the existing namespace object.

Additionally, since changing the namespace value for the component
doesn't make sense at the moment, an error is printed if a value other
than `openshift-monitoring` is detected for
`parameters.openshift4_monitoring.namepace`.

Fixes #24