Prerender `cluster-version-operator.yaml` with Jinja2
The prerendering with Jinja2 has to be done as a separate Kapitan
compile step. This is because  `kap.jinja2_template()` produces the
rendered file as YAML in contrast to `kap.yaml_load_stream()` which
produces the file contents as JSON and we cannot parse YAML in Jsonnet
(yet, cf. https://jsonnet.org/ref/stdlib.html#parseYaml).

The component now downloads the `cluster-version-operator.yaml` file
into a separate folder `manifests_requiring_prerendering` and first
renders the file with input type Jinja2.

We can use Jinja2 to prerender the file, even though it's actually
templated for Gotemplate because `{{ " \"...\" " }}` evaluates to
` "..." ` with both Jinja2 and Gotemplate.

The rules rendering in `component/rules.jsonnet` now looks for files to
process in both the `manifests/<release>` folder in the component
directory and the `prerendered_manifests/<release>` folder in the
component's folder in `compiled` (where Kapitan stores its outputs).

Finally, we add a compile step to delete the prerendered YAML again to
avoid it ending up in the cluster catalog.

Fixes #58