Add support to configure Thanos remote-write for cluster-monitoring
If enabled, setup a Thanos receiver, and configure remote-write for the
cluster-monitoring Prometheus to send all data to that receiver.

Users can supply custom configuration for the receiver (based on
kube-thanos) via parameter `thanosRemoteWrite.receiver`, and the
component will ensure that the cluster-monitoring Prometheus config is
adjusted to find the receiver regardless of customizations.

If users leave `thanosRemoteWrite.objectStorageConfig` empty, the
component will not manage the object storage config secret, and expects
that users configure the receiver to use an externally managed config
through `thanosRemoteWrite.receiver.objectStorageConfig`.