Add init container which creates bucket for Thanos receiver
We run `mc mb` for the configured bucket, iff it's an S3 bucket in an
init container of the Thanos receiver pod. This should allow users to
deploy the feature without having to manually create an S3 bucket if
they provide S3 credentials which have sufficient permissions to create
buckets.

We render an `mc` `config.json` from the provided Thanos objectstorage
config, assumihg that the S3 endpoint uses HTTPS.