Explicitly configure projected SA token volume for silence cronjob
On fresh OpenShift 4.11 installations, the `prometheus-k8s`
ServiceAccount has `automountServiceAccountToken=false` which prevents
automatic creation of the projected SA token volume. On upgraded
clusters, this configuration for the ServiceAccount isn't updated.

This commit updates the silence cronjob to have an explicit projected SA
token volume which should resolve the issue of the cronjob failing to
find the SA token on fresh 4.11 installations.