Drop low value rules by default
Where we can, we use the existing `ignoreNames` field in the hierarchy,
so the alerts could be enabled again without patching the component.

However, this doesn't work for alerts which exist with the same name and
different severities, and we want to blanket drop some classes of alerts.

To handle these alerts, we implement logic directly in Jsonnet which

* Drops all alerts with label `severity=info`
* Drops whole alert groups by group name based on the group names given
  in new parameter `ignoreGroups` (before the group name is prefixed
  with `syn-`).
* Drops selected alerts with label `severity=warning`. For this part, we
  currently keep the list of alerts to drop directly in Jsonnet.
  However, this could easily be moved into a new component parameter.