Introduce parameter which allows patching alerts for all OCP4 versions
We introduce parameter `alerts.patchRules."*"` which allows users to
patch alerts regardless of the value of `manifests_version`. This will
make managing custom alert patchsets (e.g. for marking alerts as
maintenance or on-call relevant) easier.

We also move the patches which are defined in `class/defaults.yml` to
the new parameter, since they are applicable for all currently supported
OpenShift 4 versions.

Note that it's possible to patch an alert both for all versions and for
a specific version, in which case, the patches get merged, with the
version-specific patch having higher priority.