Fix OVNKubernetes alert rules for OpenShift 4.14
OpenShift 4.14 moved the OVNKubernetes control plane alerts (and we
can't use the 4.13 alerts, because 4.14 makes significant changes to the
OVNKubernetes architecture). Additionally 4.14 introduced gotemplate
snippets in the OVNKubernetes alerts that we need to render.

Since the files also contain gotemplate snippets which aren't valid
Jinja2, we implement very simple rendering logic which uses
`std.strReplace()` in the component.