Add `openshift4-monitoring` to team-routing `applications` list
This is necessary because the projectsyn/jsonnet-libs version of
`teamForApplication()` expects that the `applications` list contains all
instances present on the cluster. Without adding `openshift4-monitoring`
the alert patching doesn't do the right thing anymore.

We add the application in a separate commit so we can properly verify
that the switch to `projectsyn/jsonnet-libs/syn-teams.libsonnet` doesn't
introduce any changes.