Fix alert routing namespace discovery logic
We tend to use `namespace: ${_instance}` for components where we deploy
each instance in a separate namespace (e.g. component-vault,
component-openshift4-operators). However, because we read the instance
namespaces from openshift4-monitoring's parameters, `${_instance}` is
resolved to `openshift4-monitoring` and not to the component instance
name.

This commit updates the discovery logic to override the discovered
namespace if we discover a namespace that contains
`openshift4-monitoring` for any app other than `openshift4-monitoring`
itself. The new override logic replaces all occurrences of
`openshift4-monitoring` in the discovered namespace with the parsed
component instance name.