Remove custom `NodeCpuLoad5` alert
We had the old version of this alert (`node_cpu_load5`) in the list of
ignored alerts (probably because we forgot that we deployed it
ourselves). Since it got renamed for v7.0.0, it didn't get dropped
anymore.

Instead of updating the ignore list, we simply remove the custom alert
rule completely.